+++ tests/prop_map.c
@@ -11,9 +11,9 @@ struct TestResult test_prop_map() {
     struct PropMap pm;
     tassert(pm_new(8, &pm) == NO_ERROR, "failed to initialize");
     tassert(pm.item_count == 0, "why are there items wtf");
-    long num = 1;
+    signed long long num = 1;
     struct Object val;
-    tassert(ctno_literal(&num, sizeof(long), NULL, &val) == NO_ERROR,
+    tassert(ctno_literal(&num, sizeof(signed long long), NULL, &val) == NO_ERROR,
             "failed to initialize val");
     struct Object *ret;
     a_set("foobar", &val);
